+++ app/src/global.js
@@ -162,6 +162,8 @@ a.ytp-ce-covering-overlay,
 a.ytp-suggestion-link,
 div.fullscreen-recommendation,
 ytd-structured-description-video-lockup-renderer,
+a.ytp-autonav-endscreen-link-container,
+a.autonav-endscreen-cued-video-container,
 
 ytm-compact-playlist-renderer,
 ytm-playlist-card-renderer` /*the last two selectors are playlist elements but are used for thumbnail*/,